point to `nanocomponent-adapters-angular`
This change also shuffles around the example into it's own dir so that
the `Button` component can be used in different lib examples. I didn't
add an Angular one because I have no idea how to use Angular and was
unsure of the work required to get it working with `bankai` seems like a
rabbit hole. We might want to just setup a seperate example project
using the `angular-cli` in the `nanocomponent-adapters-angular` repo.